Split up parser.rs
It was getting pretty big. Overall changes to other files are fairly minimal.
