+++ src/lib.rs
@@ -8,6 +8,7 @@ pub mod constraints;
 pub mod umiaq_char;
 mod joint_constraints;
 mod scan_hints;
+mod errors;
 
 // Compile the wasm glue only when targeting wasm32.
 #[cfg(target_arch = "wasm32")]
