+++ src/solver.rs
@@ -3,16 +3,14 @@ use crate::joint_constraints::{parse_joint_constraints, propagate_joint_to_var_b
 use crate::parser::{
     match_equation_all,
     parse_form,
-    ParseError,
     ParsedForm,
-    has_inlineable_var_form,
-    form_to_regex_str_with_constraints,
-    get_regex,
 };
+use crate::parser::prefilter::build_prefilter_regex;
 use crate::patterns::Patterns;
 use crate::scan_hints::{form_len_hints_pf, PatternLenHints};
 
 use std::collections::{HashMap, HashSet};
+use crate::errors::ParseError;
 
 /// The max number of matches to grab during our initial pass through the word list
 const MAX_INITIAL_MATCHES: usize = 500_000;
@@ -250,19 +248,10 @@ pub fn solve_equation(input: &str, word_list: &[&str], num_results_requested: us
     let mut var_constraints = patterns.var_constraints.clone();
 
     // 4a. Upgrade prefilters once per form (only if it helps)
+    // parsed_forms is mutable
     for pf in &mut parsed_forms {
-        if has_inlineable_var_form(&pf.parts, &var_constraints) {
-            // Build the anchored, constraint-aware pattern string
-            let anchored = format!(
-                "^{}$",
-                form_to_regex_str_with_constraints(&pf.parts, Some(&var_constraints))
-            );
-
-            // Compile via the shared cache; fall back to the existing prefilter on error
-            if let Ok(re) = get_regex(&anchored) {
-                pf.prefilter = re;
-            }
-        }
+        let upgraded = build_prefilter_regex(pf, Some(&var_constraints));
+        pf.prefilter = upgraded;
     }
 
     // 4b. Get the joint constraints
